Güvenlik güncellemesi: admin erişimi server-side auth ile güvenli hale getirildi
