+++ app/admin/components/UsersTable.tsx
@@ -1,0 +1,97 @@
+//app/admin/components/UsersTable.tsx
+
+'use client';
+
+import { useEffect, useState } from 'react';
+
+type User = {
+  id: string;
+  name: string | null;
+  email: string | null;
+  role: string;
+};
+
+export default function UsersTable() {
+  const [users, setUsers] = useState<User[]>([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const fetchUsers = async () => {
+    setLoading(true);
+    setError(null);
+    try {
+      const res = await fetch('/api/admin/users');
+      if (!res.ok) {
+        const data = await res.json();
+        throw new Error(data.message || 'Failed to fetch users');
+      }
+      const data = await res.json();
+      setUsers(data);
+    } catch (err: any) {
+      setError(err.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleRoleChange = async (userId: string, newRole: string) => {
+    try {
+      const res = await fetch('/api/admin/users', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ userId, newRole }),
+      });
+
+      if (!res.ok) {
+        const data = await res.json();
+        throw new Error(data.message || 'Failed to update role');
+      }
+      fetchUsers();
+    } catch (err: any) {
+      alert(err.message);
+    }
+  };
+
+  useEffect(() => {
+    fetchUsers();
+  }, []);
+
+  if (loading) return <p>Loading users...</p>;
+  if (error) return <p className="text-red-500">Error: {error}</p>;
+
+  return (
+    <table className="min-w-full bg-white border border-gray-300">
+      <thead>
+        <tr className="bg-gray-100">
+          <th className="py-2 px-4 border-b">ID</th>
+          <th className="py-2 px-4 border-b">Name</th>
+          <th className="py-2 px-4 border-b">Email</th>
+          <th className="py-2 px-4 border-b">Role</th>
+          <th className="py-2 px-4 border-b">Actions</th>
+        </tr>
+      </thead>
+      <tbody>
+        {users.map((user) => (
+          <tr key={user.id}>
+            <td className="py-2 px-4 border-b">{user.id}</td>
+            <td className="py-2 px-4 border-b">{user.name}</td>
+            <td className="py-2 px-4 border-b">{user.email}</td>
+            <td className="py-2 px-4 border-b">{user.role}</td>
+            <td className="py-2 px-4 border-b">
+              <select
+                value={user.role}
+                onChange={(e) => handleRoleChange(user.id, e.target.value)}
+                className="border rounded px-2 py-1"
+              >
+                <option value="SUPER_ADMIN">SUPER_ADMIN</option>
+                <option value="ADMIN">ADMIN</option>
+                <option value="DEVELOPER">DEVELOPER</option>
+                <option value="USER">USER</option>
+              </select>
+            </td>
+          </tr>
+        ))}
+      </tbody>
+    </table>
+  );
+}
