+++ app/admin/page.tsx
@@ -1,118 +1,20 @@
-// 📄 app/admin/page.tsx
-'use client';
+//app/admin/page.tsx
 
-import { useEffect, useState } from 'react';
-import { useSession } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
+import { auth } from '@lib/auth';
+import { redirect } from 'next/navigation';
+import UsersTable from './components/UsersTable';
 
-type User = {
-  id: string;
-  name: string | null;
-  email: string | null;
-  role: string;
-};
+export default async function AdminPage() {
+  const session = await auth();
 
-export default function AdminPage() {
-  const { data: session, status } = useSession();
-  const router = useRouter();
-  const [users, setUsers] = useState<User[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    // ✅ Eğer kullanıcı giriş yapmamışsa veya SUPER_ADMIN değilse yönlendir
-    if (status === 'loading') return;
-    if (!session || session.user.role !== 'SUPER_ADMIN') {
-      router.replace('/signin'); 
-    }
-  }, [session, status, router]);
-
-  const fetchUsers = async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const res = await fetch('/api/admin/users');
-      if (!res.ok) {
-        const data = await res.json();
-        throw new Error(data.message || 'Failed to fetch users');
-      }
-      const data = await res.json();
-      setUsers(data);
-    } catch (err: any) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleRoleChange = async (userId: string, newRole: string) => {
-    try {
-      const res = await fetch('/api/admin/users', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ userId, newRole }),
-      });
-
-      if (!res.ok) {
-        const data = await res.json();
-        throw new Error(data.message || 'Failed to update role');
-      }
-      fetchUsers();
-    } catch (err: any) {
-      alert(err.message);
-    }
-  };
-
-  useEffect(() => {
-    if (session?.user?.role === 'SUPER_ADMIN') {
-      fetchUsers();
-    }
-  }, [session]);
-  if (status === 'loading') return <p>Loading session...</p>;
+  if (!session || session.user.role !== 'SUPER_ADMIN') {
+    redirect('/signin');
+  }
 
   return (
     <div className="p-8">
       <h2 className="text-xl font-bold mb-4">Users Table</h2>
-      {loading && <p>Loading users...</p>}
-      {error && <p className="text-red-500">Error: {error}</p>}
-      {!loading && !error && (
-        <table className="min-w-full bg-white border border-gray-300">
-          <thead>
-            <tr className="bg-gray-100">
-              <th className="py-2 px-4 border-b">ID</th>
-              <th className="py-2 px-4 border-b">Name</th>
-              <th className="py-2 px-4 border-b">Email</th>
-              <th className="py-2 px-4 border-b">Role</th>
-              <th className="py-2 px-4 border-b">Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            {users.map((user) => (
-              <tr key={user.id}>
-                <td className="py-2 px-4 border-b">{user.id}</td>
-                <td className="py-2 px-4 border-b">{user.name}</td>
-                <td className="py-2 px-4 border-b">{user.email}</td>
-                <td className="py-2 px-4 border-b">{user.role}</td>
-                <td className="py-2 px-4 border-b">
-                  <select
-                    value={user.role}
-                    onChange={(e) => handleRoleChange(user.id, e.target.value)}
-                    className="border rounded px-2 py-1"
-                  >
-                    <option value="SUPER_ADMIN">SUPER_ADMIN</option>
-                    <option value="ADMIN">ADMIN</option>
-                    <option value="DEVELOPER">DEVELOPER</option>
-                    <option value="USER">USER</option>
-                  </select>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      )}
+      <UsersTable />
     </div>
   );
 }
-
-
-// 📌 Not: Admin sayfasında artık genel navbar veya product topbar kullanılmıyor. Admin'e özel sade bir sayfa tasarlandı.
